alle IndexOfObject zu long geändert: int erzeugte falsche Angaben für NSNotFound Speaker: Op-sound ok Testliste OK: NameZuTestDS und TestZuNamenDS hatten keine tableview-delegates eingetragen NamenZuTestCheckboxAktion hatte falschen Aufruf von tempNamenDic: cast zu rNamenZuTestDS*, Aufruf von NamenZuTestDicArray. TestZuNamenCheckboxAktion dto

+++ SndCalcX/rNamenListe.m
@@ -159,7 +159,7 @@ extern NSString* OK;
 		if(EntfernenName)
 		{
 			NSArray* tempNamenArray=[NamenArray valueForKey:@"namen"];
-			int deleteIndex=[tempNamenArray indexOfObject:EntfernenName];
+			long deleteIndex=[tempNamenArray indexOfObject:EntfernenName];
 			
 			if (deleteIndex>=0)//EntfernenName ist vorhanden
 			{
